Delete legacy create_database_from_conn_id
