+++ tests/utils/test_database.py
@@ -7,28 +7,9 @@ from sqlalchemy import text
 from sqlalchemy.engine.base import Engine
 
 from astro.constants import Database
-from astro.utils.database import (
-    create_database_from_conn_id,
-    get_database_name,
-    get_sqlalchemy_engine,
-    run_sql,
-)
+from astro.utils.database import get_database_name, get_sqlalchemy_engine, run_sql
 from astro.utils.dependencies import BigQueryHook, PostgresHook, SnowflakeHook
 from astro.utils.sqlite_utils import create_sqlalchemy_engine_with_sqlite
-
-
-def describe_create_database():
-    def with_supported_databases(session):
-        assert create_database_from_conn_id("postgres_default") == Database.POSTGRES
-        assert create_database_from_conn_id("sqlite_default") == Database.SQLITE
-        assert create_database_from_conn_id("google_cloud_default") == Database.BIGQUERY
-        assert create_database_from_conn_id("snowflake_conn") == Database.SNOWFLAKE
-
-    def with_unsupported_database(session):
-        with pytest.raises(ValueError) as exc_info:
-            assert create_database_from_conn_id("cassandra_default")
-        expected_msg = "Unsupported database <cassandra>"
-        assert exc_info.value.args[0] == expected_msg
 
 
 def describe_get_database_name():
